update: now Win64-build with VisualStudio2010
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2070 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/basis/modpoly/algorithm_sweep/DehierarchisationModPoly.cpp
@@ -46,7 +46,7 @@ void DehierarchisationModPoly::rec(DataVector& source, DataVector& result, grid_
 	index.get(dim, cur_lev, cur_ind);
 		
 	// Dehierarchisation
-	result[seq] = source[seq] + this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, cur_lev)));
+	result[seq] = source[seq] + this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, static_cast<int>(cur_lev))));
 	
 	// recursive calls for the right and left side of the current node
 	if(index.hint() == false)
