+++ src/sgpp/base/basis/modpoly/algorithm_sweep/HierarchisationModPoly.cpp
@@ -45,7 +45,7 @@ void HierarchisationModPoly::rec(DataVector& source, DataVector& result, grid_it
 	index.get(dim, cur_lev, cur_ind);
 		
 	// hierarchisation
-	result[seq] = source[seq] - this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, cur_lev)));
+	result[seq] = source[seq] - this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, static_cast<int>(cur_lev))));
 	
 	// recursive calls for the right and left side of the current node
 	if(index.hint() == false)
