+++ src/sgpp/base/basis/poly/algorithm_sweep/HierarchisationPoly.cpp
@@ -45,7 +45,7 @@ void HierarchisationPoly::rec(DataVector& source, DataVector& result, grid_itera
 	index.get(dim, cur_lev, cur_ind);
 		
 	// hierarchisation
-	result[seq] = source[seq] - this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, cur_lev)));
+	result[seq] = source[seq] - this->base->evalHierToTop(cur_lev, cur_ind, koeffs, cur_ind/(pow(2.0, static_cast<int>(cur_lev))));
 	
 	// recursive calls for the right and left side of the current node
 	if(index.hint() == false)
