+++ src/sgpp/base/grid/generation/SurplusVolumeRefinementFunctor.cpp
@@ -24,7 +24,7 @@ SurplusVolumeRefinementFunctor::~SurplusVolumeRefinementFunctor()
 
 double SurplusVolumeRefinementFunctor::operator()(GridStorage* storage, size_t seq)
 {
-  return pow(2, - ((int)storage->get(seq)->getLevelSum()))*fabs(alpha->get(seq));
+  return pow(2, static_cast<double>(-((int)storage->get(seq)->getLevelSum()))*fabs(alpha->get(seq)));
 }
 
 double SurplusVolumeRefinementFunctor::start()
